requirements.txt
gpiozero>=2.0
RPi.GPIO>=0.7.1
smbus2>=0.4.3
python-rtmidi>=1.5.8
construct>=2.10.70
netifaces>=0.11.0
pyudev>=0.24.1
